Resolved register issue
