+++ src/components/button/index.tsx
@@ -1,5 +1,1 @@
-import Button from  './Button.component';
-export default Button;
-import BR from '@blueeast/bluerain-os';
-
-BR.Components.register('Button', Button);
+export { default } from  './Button.component';
