+++ src/index.tsx
@@ -5,7 +5,8 @@ import TextInput from './components/textInput';
 import Chip from './components/chip';
 import GMap from './components/map';
 import Badge from './components/badge';
-
+import Button from './components/button';
+import Card, {CardHeader, CardBody, CardFooter, CardMedia} from './components/Card';
 // import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
 // import getMuiTheme from 'material-ui/styles/withTheme';
 
@@ -44,6 +45,12 @@ class MaterialUiPlugin extends Plugin {
         ctx.Components.register('Chip', Chip);
         ctx.Components.register('Badge', Badge);
         ctx.Components.register('GMap', GMap);
+        ctx.Components.register('Button', Button);
+        ctx.Components.register('Card', Card);
+        ctx.Components.register('CardHeader', CardHeader);
+        ctx.Components.register('CardFooter', CardFooter);
+        ctx.Components.register('CardBody', CardBody);
+        ctx.Components.register('CardMedia', CardMedia);
 
         // Add Material UI Provider
 //         	ctx.Filters.add(
